requirements.txt
pydantic-ai-slim[openai,anthropic]
